rgw: move rgw_sync_log_trim.* to rgw_trim_bilog.*
