+++ src/rgw/CMakeLists.txt
@@ -72,8 +72,8 @@ set(librgw_common_srcs
   rgw_sync_module_pubsub.cc
   rgw_pubsub_push.cc
   rgw_sync_module_pubsub_rest.cc
-  rgw_sync_log_trim.cc
   rgw_sync_trace.cc
+  rgw_trim_bilog.cc
   rgw_trim_datalog.cc
   rgw_trim_mdlog.cc
   rgw_period_history.cc
